Weekend save

+++ dice-calculator/struct_sortedinput.h
@@ -1,8 +1,10 @@
 #pragma once
 #include <vector>
-#include "Diceset.h"
+#include "class_Diceset.h"
+
+// The structure sortedinput is used to pass the getdicesets return values to the other functions
 
 struct sortedinput {
-    int addsubtract{};
+    int64_t addsubtract{};
     std::vector<Diceset> dicesets;
 } typedef sortedinput;
